% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callGSVA.R
\name{callGSVA}
\alias{callGSVA}
\title{GSVA enrichment analysis}
\usage{
callGSVA(x,y)
}
\arguments{
\item{x}{A data frame or matrix of gene or probe expression values where rows corrospond to genes and columns corrospond to samples}

\item{y}{A list of genes as data frame or vector}
}
\value{
A gene-set by sample matrix of GSVA enrichment zscores.
}
\description{
Estimates GSVA enrichment zscores.
}
\details{
This function uses "zscore" gene-set enrichment method in the estimation of gene-set enrichment scores per sample.
}
\examples{
g <- 10 ## number of genes
s <- 30 ## number of samples
## sample data matrix with values ranging from 1 to 10
rnames <- paste("g", 1:g, sep="")
cnames <- paste("s", 1:s, sep="")
expr <- matrix(sample.int(10, size = g*s, replace = TRUE), nrow=g, ncol=s, dimnames=list(rnames, cnames))
## genes of interest
genes <- data.frame(paste("g", 1:6, sep=""))
## Estimates GSVA enrichment zscores.
callGSVA(expr,genes)
}
\seealso{
GSVA
}

