% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callZSCORE.R
\name{callZSCORE}
\alias{callZSCORE}
\title{Row ZSCORES}
\usage{
callZSCORE(x)
}
\arguments{
\item{x}{A data frame or matrix of gene or probe expression values where rows corrospond to genes and columns corrospond to samples}
}
\value{
A gene-set by sample matrix of zscores.
}
\description{
Estimates the zscores for each row in the data matrix
}
\details{
This function compute row zscores per sample when number of genes is less than 3
}
\examples{
g <- 2 ## number of genes
s <- 60 ## number of samples
## sample data matrix with values ranging from 1 to 10
rnames <- paste("g", 1:g, sep="")
cnames <- paste("s", 1:s, sep="")
expr <- matrix(sample.int(10, size = g*s, replace = TRUE), nrow=g, ncol=s, dimnames=list(rnames, cnames))
## Estimates zscores
callZSCORE(expr)
}
