% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/changepoints_all.R
\docType{data}
\name{changepoints_all}
\alias{changepoints_all}
\title{An example input data frame with added sample identifiers and estimated changepoints}
\usage{
data(changepoints_all)
}
\value{
data frame consisting of samples, zscores, converted zscores, identified changepoints, and assigned sample groups
}
\description{
The input data frame with with 125 sample rows and 4 columns:
}
\details{
This is data to be included in my package
}
\keyword{datasets}

