% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression_data.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{An example of RNAseq derived gene expression data}
\usage{
data(expression_data)
}
\value{
numeric expression dataset of 8 probes (rows) on 125 samples (column)
}
\description{
This dataset contains the expression values of 8 probes (rows) in 125 samples (columns), as compiled by the CoMMpass study.
}
\details{
This is data to be included in my package
}
\keyword{datasets}

