% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExpressionSet}
\alias{ExpressionSet}
\title{Gene Expression (ExpressionSet) data}
\format{
A data matrix with 8 probes (rows) in 125 samples (columns)
}
\usage{
ExpressionSet
}
\description{
An example of RNAseq derived gene expression data
The variables are as follows
}
\details{
\itemize{
 \item gene
 \item sample 1 log2 transformed normalized expression count values
 \item sample nth log2 transformed normalized expression count values
 }
 @source \url{https://research.themmrf.org/
 }
}
\keyword{datasets}
