% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVars.R
\name{filterVars}
\alias{filterVars}
\title{A filter function for the data}
\usage{
filterVars(x,y)
}
\arguments{
\item{x}{: A data frame or matrix where rows represent gene and columns represent samples}

\item{y}{: A vector of a sample column values to apply the filtering on.}
}
\value{
The returned value is a list containing an entry for each row filtered out by zero data value
}
\description{
Filter rows with zero values
}
\details{
This function filter out rows with zero data value for a given sample. Both input arguments (x and y) must be of the same length
}
\examples{
x = matrix(runif(3*10, 0, 1), ncol=3)
y <- x[,1]
filterVars(x,y)
}
