% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freqplot.R
\name{freqplot}
\alias{freqplot}
\title{A plotting function for SISPA sample identifiers}
\usage{
freqplot(x)
}
\arguments{
\item{x}{: A data frame containing samples as rows followed by zscores and estimated changepoints to be plotted.}
}
\value{
Bar plot pdf illustrating distribution of samples
}
\description{
Given a sample changepoint data frame, will plot number of samples with and without profile activity
}
\details{
This function expects the output from cptSamples function of SISPA package, and shows the number of samples with (orange filled bars) and without profile activity (grey filled bars).
}
\examples{
g <- 10 ## number of genes
s <- 30 ## number of samples
## sample data matrix with values ranging from 1 to 10
expr <- matrix(sample.int(10, size = g*s, replace = TRUE), nrow=g, ncol=s, dimnames=list(paste("g", 1:g, sep="") , paste("s", 1:s, sep="")))
## genes of interest
genes <- list(set1=paste("g", 1:3, sep=""))
## Estimates GSVA enrichment zscores.
gsva_results <- callGSVA(expr,genes)
cpt_on_samples <- cptSamples(gsva_results,dir="up",cpt_data="var",cpt_method="BinSeg",cpt_max=60)
## Plot number of samples by sample groups
freqplot(cpt_on_samples)
}

