% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gene_list.R
\docType{data}
\name{gene_list}
\alias{gene_list}
\title{an example file containing the list of input genes}
\usage{
data(gene_list)
}
\value{
list of 8 genes(probes)
}
\description{
A list of 9 genes for GSVA analysis
}
\details{
This is data to be included in my package
}
\keyword{datasets}

