% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gsva_results.R
\docType{data}
\name{gsva_results}
\alias{gsva_results}
\title{An example output from sample callGSVA function}
\usage{
data(gsva_results)
}
\value{
a data matrix consisting of 125 samples zscores
}
\description{
A gene-set by sample matrix consisting of 125 sample GSVA enrichment zscores
}
\details{
This is data to be included in my package
}
\keyword{datasets}

