% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortData.R
\name{sortData}
\alias{sortData}
\title{Sorts the data by a column}
\usage{
sortData(x,i,b)
}
\arguments{
\item{x}{A data frame}

\item{i}{A numeric column index of the data frame to sort it by}

\item{b}{User specified sorting order, ascending (FALSE) or descending (TRUE)}
}
\value{
sorted data by the input column index
}
\description{
Sorts the data frame by a column index in the given order
}
\details{
defaults are used: i = 1, b = FALSE, if not specified
}
\examples{
samples <- c("s1","s2","s3","s4","s5","s6","s7","s8","s9","s10")
zscores <- c(3.83,2.70,2.67,2.31,1.70,1.25,-0.42,-1.01,-2.43,-3.37)
my.data = data.frame(samples,zscores)
sortData(my.data,2,TRUE)
}
\author{
Bhakti Dwivedi & Jeanne Kowalski
}
