% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variant_data.R
\docType{data}
\name{variant_data}
\alias{variant_data}
\title{An example of RNAseq derived gene variant change data}
\usage{
data(variant_data)
}
\value{
numeric variant dataset of 380 variants (rows) on 125 samples (column)
}
\description{
This dataset contains the variant proportion values of variants (n=380) associated with 8 genes (rows) in 125 samples (columns), as compiled by the CoMMpass study.
}
\details{
This is data to be included in my package
}
\keyword{datasets}

