% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variant_list.R
\docType{data}
\name{variant_list}
\alias{variant_list}
\title{an example file containing the list of input genes variants}
\usage{
data(variant_list)
}
\value{
list of 380 gene variants
}
\description{
A list of variants associated with 8 genes for GSVA analysis
}
\details{
This is data to be included in my package
}
\keyword{datasets}

