% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VariantSet}
\alias{VariantSet}
\title{Variant (VariantSet) data}
\format{
A data matrix with 380 variants (rows) in 125 samples (column)
}
\usage{
VariantSet
}
\description{
An example of RNAseq derived gene variant change data.
The variables are as follows
}
\details{
\itemize{
 \item gene
 \item sample 1 transformed variant proportion data
 \item sample nth transformed variant proportion data
 }
 @source \url{https://research.themmrf.org/
 }
}
\keyword{datasets}
