% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterfallplot.R
\name{waterfallplot}
\alias{waterfallplot}
\title{A plotting function for SISPA sample identifiers}
\usage{
waterfallplot(x)
}
\arguments{
\item{x}{A data frame containing samples as rows followed by zscores and estimated sample_groups to be plotted.}
}
\value{
Bar plot pdf illustrating distinct SISPA sample profiles.
}
\description{
Given a sample changepoint data frame, will plot all samples zscores from that data.
}
\details{
This function expects the output from cptSamples function of SISPA package, and highlights the sample profile of interest in the changepoint 1 with orange filled bars.
}
\examples{
samples <- c("s1","s2","s3","s4","s5","s6","s7","s8","s9","s10")
zscores <- c(3.83,2.70,2.67,2.31,1.70,1.25,-0.42,-1.01,-2.43,-3.37)
changepoints <- c(1,1,1,2,2,3,3,NA,NA,NA)
sample_groups <- c(1,1,1,0,0,0,0,0,0,0)
my.data = data.frame(samples,zscores,changepoints,sample_groups)
waterfallplot(my.data)
}

