% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callGSVA.R
\name{callGSVA}
\alias{callGSVA}
\title{GSVA enrichment analysis}
\usage{
callGSVA(x,y)
}
\arguments{
\item{x}{: A data matrix of gene or probe expression values where rows corrospond to genes and columns corrospond to samples}

\item{y}{: Gene sets provided as a list object.}
}
\value{
A gene-set by sample matrix of GSVA enrichment zscores.
}
\description{
Estimates GSVA enrichment zscores.
}
\details{
This function uses "zscore" gene-set enrichment method in the estimation of gene-set enrichment scores per sample.
}
\examples{
g <- 10 ## number of genes
s <- 30 ## number of samples
## sample data matrix with values ranging from 1 to 10
expr <- matrix(sample.int(10, size = g*s, replace = TRUE), nrow=g, ncol=s, dimnames=list(paste("g", 1:g, sep="") , paste("s", 1:s, sep="")))
## genes of interest
genes <- list(set1=paste("g", 1:3, sep=""))
## Estimates GSVA enrichment zscores.
callGSVA(expr,genes)
}
\seealso{
GSVA
}

