% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SISPA.R
\name{SISPA}
\alias{SISPA}
\title{SISPA}
\usage{
SISPA(x,y)
}
\arguments{
\item{x}{: A data matrix of gene or probe expression values where rows corrospond to genes and columns corrospond to samples}

\item{y}{: Gene sets provided as a list object.}
}
\value{
The input molecular data frame with added sample identifiers and estimated changepoints. A plot showing the changepoint locations estimated on the data. Bar plots pdf illustrating distinct distribution of samples with and without profile activity
}
\description{
SISPA: Method for Sample Integrated Gene Set Analysis
}
\details{
Sample Integrated Gene Set Analysis (SISPA) is a method designed to define sample groups with similar gene set enrichment profiles. The user specifies a gene list of interest and sample by gene molecular data (expression, methylation, variant, or copy change data) to obtain gene set enrichment scores by each sample. The score statistics is rank ordered by the desired profile (e.g., upregulated or downregulated) for samples. A change point model is then applied to the sample scores to identify groups of samples that show similar gene set profile patterns. Samples are ranked by desired profile activity score and grouped by samples with and without profile activity. Figure 1 shows the schematic representation of the SISPA method overview.
}
\examples{
g <- 10 ## number of genes
s <- 60 ## number of samples
## sample data matrix with values ranging from 1 to 10
expr <- matrix(sample.int(10, size = g*s, replace = TRUE), nrow=g, ncol=s, dimnames=list(paste("g", 1:g, sep="") , paste("s", 1:s, sep="")))
## genes of interest
genes <- list(set1=paste("g", 1:6, sep=""))
SISPA(expr,genes)
}

