% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterfallplot.R
\name{waterfallplot}
\alias{waterfallplot}
\title{A plotting function for SISPA sample identifiers}
\usage{
waterfallplot(x)
}
\arguments{
\item{x}{: A data frame containing samples as rows followed by zscores and estimated changepoints to be plotted.}
}
\value{
Bar plot pdf illustrating distinct SISPA sample profiles.
}
\description{
Given a sample changepoint data frame, will plot all samples zscores from that data.
}
\details{
This function expects the output from cptSamples function of SISPA package, and highlights the sample profile of interest in the changepoint 1 with orange filled bars.
}
\examples{
g <- 50 ## number of genes
s <- 50 ## number of samples
## sample data matrix with values ranging from 1 to 50
expr <- matrix(sample.int(50, size = g*s, replace = TRUE), nrow=g, ncol=s, dimnames=list(paste("g", 1:s, sep="") , paste("s", 1:g, sep="")))
## genes of interest
genes <- list(set1=paste("g", 1:3, sep=""))
## Estimates GSVA enrichment zscores.
gsva_results <- callGSVA(expr,genes)
cpt_on_samples <- cptSamples(gsva_results,dir="up",cpt_data="var",cpt_method="BinSeg",cpt_max=60)
waterfallplot(cpt_on_samples)
}

